+++ packages/addons-v5/commands/addons/open.js
@@ -3,13 +3,15 @@
 let cli = require('heroku-cli-util')
 let co = require('co')
 let fs = require('fs')
+let os = require('os')
+let path = require('path')
 
 function open (url) {
   cli.log(`Opening ${cli.color.cyan(url)}...`)
   return cli.open(url)
 }
 
-const ssoPath = '/tmp/heroku-sso.html'
+const ssoPath = path.join(os.tmpdir(), 'heroku-sso.html')
 
 function writeSudoTemplate (ctx, sso, path) {
   return new Promise(function (resolve, reject) {