+++ packages/run-v5/test/commands/logs.js
@@ -8,7 +8,7 @@ describe('logs', () => {
   beforeEach(() => cli.mockConsole())
 
   it('shows the logs', () => {
-    return cmd.run({ app: 'heroku-run-test-app', flags: {}, auth: { password: global.apikey } })
+    return cmd.run({ app: 'heroku-cli-ci-smoke-test-app', flags: {}, auth: { password: global.apikey } })
       .then(() => expect(cli.stdout, 'to begin with', '20')) // starts with the year
   })
 })