+++ packages/run-v5/test/commands/run.js
@@ -13,7 +13,7 @@ describe('run', () => {
     let stdout = ''
     let fixture = new StdOutFixture()
     fixture.capture(s => { stdout += s })
-    return cmd.run({ app: 'heroku-run-test-app', flags: {}, auth: { password: global.apikey }, args: ['echo', '1', '2', '3'] })
+    return cmd.run({ app: 'heroku-cli-ci-smoke-test-app', flags: {}, auth: { password: global.apikey }, args: ['echo', '1', '2', '3'] })
       .then(() => fixture.release())
       .then(() => expect(stdout, 'to contain', '1 2 3\n'))
   })
@@ -22,7 +22,7 @@ describe('run', () => {
     let stdout = ''
     let fixture = new StdOutFixture()
     fixture.capture(s => { stdout += s })
-    return cmd.run({ app: 'heroku-run-test-app', flags: {}, auth: { password: global.apikey }, args: ['ruby', '-e', 'puts ARGV[0]', '{"foo": "bar"} '] })
+    return cmd.run({ app: 'heroku-cli-ci-smoke-test-app', flags: {}, auth: { password: global.apikey }, args: ['ruby', '-e', 'puts ARGV[0]', '{"foo": "bar"} '] })
       .then(() => fixture.release())
       .then(() => expect(stdout, 'to equal', '{"foo": "bar"} \n'))
   })
@@ -31,7 +31,7 @@ describe('run', () => {
     let stdout = ''
     let fixture = new StdOutFixture()
     fixture.capture(s => { stdout += s })
-    return cmd.run({ app: 'heroku-run-test-app', flags: {}, auth: { password: global.apikey }, args: ['ruby', '-e', 'puts ARGV[0]', '{"foo":"bar"}'] })
+    return cmd.run({ app: 'heroku-cli-ci-smoke-test-app', flags: {}, auth: { password: global.apikey }, args: ['ruby', '-e', 'puts ARGV[0]', '{"foo":"bar"}'] })
       .then(() => fixture.release())
       .then(() => expect(stdout, 'to equal', '{"foo":"bar"}\n'))
   })
@@ -40,7 +40,7 @@ describe('run', () => {
     let stdout = ''
     let fixture = new StdOutFixture()
     fixture.capture(s => { stdout += s })
-    return cmd.run({ app: 'heroku-run-test-app', flags: { env: 'FOO=bar' }, auth: { password: global.apikey }, args: ['env'] })
+    return cmd.run({ app: 'heroku-cli-ci-smoke-test-app', flags: { env: 'FOO=bar' }, auth: { password: global.apikey }, args: ['env'] })
       .then(() => fixture.release())
       .then(() => expect(stdout, 'to contain', 'FOO=bar'))
   })
@@ -48,7 +48,7 @@ describe('run', () => {
   it('gets 127 status for invalid command', () => {
     cli.exit.mock()
     return assertExit(127, cmd.run({
-      app: 'heroku-run-test-app',
+      app: 'heroku-cli-ci-smoke-test-app',
       flags: { 'exit-code': true },
       auth: { password: global.apikey },
       args: ['invalid-command'] }))