+++ packages/run/test/commands/logs.test.ts
@@ -3,7 +3,7 @@ import {expect, test} from '@oclif/test'
 describe('logs', () => {
   test
   .stdout()
-  .command(['logs', '--app=heroku-run-test-app'])
+  .command(['logs', '--app=heroku-cli-ci-smoke-test-app'])
   .it('shows the logs', ctx => {
     // This is asserting that logs are returned by checking for the presence of the first two
     // digits of the year in the timetstamp