+++ packages/spaces/lib/format.js
@@ -14,52 +14,76 @@ module.exports = function () {
   }
 
   function VPNStatus (s) {
-    let colored = s
     switch (s) {
       case 'UP':
       case 'available':
-        colored = `${cli.color.green(colored)}`
-        break
+        return `${cli.color.green(s)}`
       case 'pending':
       case 'provisioning':
       case 'deprovisioning':
-        colored = `${cli.color.yellow(colored)}`
-        break
+        return `${cli.color.yellow(s)}`
       case 'DOWN':
       case 'deleting':
       case 'deleted':
-        colored = `${cli.color.red(colored)}`
-        break
+        return `${cli.color.red(s)}`
+      default:
+        return s
     }
-
-    return colored
   }
 
   function PeeringStatus (s) {
-    var colored = s
     switch (s) {
       case 'active':
-        colored = `${cli.color.green(colored)}`
-        break
+        return `${cli.color.green(s)}`
       case 'pending-acceptance':
       case 'provisioning':
-        colored = `${cli.color.yellow(colored)}`
-        break
+        return `${cli.color.yellow(s)}`
       case 'expired':
       case 'failed':
       case 'deleted':
       case 'rejected':
-        colored = `${cli.color.red(colored)}`
-        break
+        return `${cli.color.red(s)}`
+      default:
+        return s
     }
+  }
 
-    return colored
+  function HostStatus (s) {
+    switch (s) {
+      case 'available':
+        return `${cli.color.green(s)}`
+      case 'under-assessment':
+        return `${cli.color.yellow(s)}`
+      case 'permanent-failure':
+      case 'released-permanent-failure':
+        return `${cli.color.red(s)}`
+      case 'released':
+        return `${cli.color.gray(s)}`
+      default:
+        return s
+    }
+  }
+
+  function Percent (v) {
+    const fmt = () => `${v}%`
+
+    switch (typeof v) {
+      case 'number':
+      case 'bigint':
+        return fmt()
+      case 'string':
+        return v.length > 0 ? fmt() : v
+      default:
+        return v
+    }
   }
 
   return {
     CIDR,
     CIDRBlocksOrCIDRBlock,
     PeeringStatus,
-    VPNStatus
+    VPNStatus,
+    HostStatus,
+    Percent
   }
 }